+++ CHANGELOG.md
@@ -6,11 +6,24 @@ HEAD
 
 * Added `BasicJsonDocument::shrinkToFit()`
 * Added support of `uint8_t` for `serializeJson()`, `serializeJsonPretty()`, and `serializeMsgPack()` (issue #1142)
+* Added `ARDUINOJSON_ENABLE_COMMENTS` to enable support for comments (defaults to 0)
 * Auto enable support for `std::string` and `std::stream` on modern compilers (issue #1156)
   (No need to define `ARDUINOJSON_ENABLE_STD_STRING` and `ARDUINOJSON_ENABLE_STD_STREAM` anymore)
 * Improved decoding of UTF-16 surrogate pairs (PR #1157 by @kaysievers)
   (ArduinoJson now produces standard UTF-8 instead of CESU-8)
 
+> ### BREAKING CHANGES
+> 
+> #### Comments
+> 
+> Support for comments in input is now optional and disabled by default.
+>
+> If you need support for comments, you must defined `ARDUINOJSON_ENABLE_COMMENTS` to `1`; otherwise, you'll receive `InvalidInput` errors.
+>
+> ```c++
+> #define ARDUINOJSON_ENABLE_COMMENTS 1
+> #include <ArduinoJson.h>
+> ```
 
 v6.13.0 (2019-11-01)
 -------